+++ src/main/java/com/solta/email/dto/request/AuthCodeEmailDTO.java
@@ -1,9 +1,10 @@
-package com.solta.email.dto;
+package com.solta.email.dto.request;
 
 import jakarta.validation.constraints.Pattern;
 
-public record EmailDTO(
+public record AuthCodeEmailDTO(
         @Pattern(regexp = "^[a-zA-Z0-9+-\\_.]+@[a-zA-Z0-9-]+\\.[a-zA-Z0-9-.]+$", message = "올바른 이메일 형식이 아닙니다.")
-        String email
+        String email,
+        String authCode
 ) {
 }
