+++ src/main/java/com/solta/email/dto/request/EmailDTO.java
@@ -1,10 +1,9 @@
-package com.solta.email.dto;
+package com.solta.email.dto.request;
 
 import jakarta.validation.constraints.Pattern;
 
-public record AuthCodeEmailDTO(
+public record EmailDTO(
         @Pattern(regexp = "^[a-zA-Z0-9+-\\_.]+@[a-zA-Z0-9-]+\\.[a-zA-Z0-9-.]+$", message = "올바른 이메일 형식이 아닙니다.")
-        String email,
-        String authCode
+        String email
 ) {
 }
